+++ src/module/applications/apps/ability-configuration-dialog.mjs
@@ -6,6 +6,7 @@ import { systemPath } from "../../constants.mjs";
  * @import AbilityModel from "../../data/item/ability.mjs";
  * @import { DrawSteelActor, DrawSteelItem } from "../../documents/_module.mjs";
  * @import DrawSteelToken  from "../../canvas/placeables/token.mjs"
+ * @import RegionDocument from "@client/documents/region.mjs";
  */
 
 /**
@@ -17,6 +18,7 @@ export default class AbilityConfigurationDialog extends PowerRollDialog {
     classes: ["ability-configuration-dialog"],
     actions: {
       panToken: this.#panToken,
+      placeTemplate: this.#placeTemplate,
     },
   };
 
@@ -178,17 +180,17 @@ export default class AbilityConfigurationDialog extends PowerRollDialog {
     context.resource.show = this.item.system.resource;
 
     // Heroic resource/malice spend
-    if (this.item.system.spend.value || this.item.system.spend.text) {
+    if (this.item.system.effects.documentsByType.spend.length) {
       context.resource.show = true;
       const coreResource = this.actor.system.coreResource;
+      context.resource.name = coreResource.name;
+      context.resource.max = foundry.utils.getProperty(coreResource.target, coreResource.path) - coreResource.minimum;
 
-      const max = foundry.utils.getProperty(coreResource.target, coreResource.path) - coreResource.minimum;
-      if (max) context.spendConfig = {
-        max,
-        slider: !this.item.system.spend.value,
-        value: this.item.system.spend.value || "",
-        name: coreResource.name,
-      };
+      context.spendConfig = this.item.system.effects.documentsByType.spend.map(e => ({
+        id: e.id,
+        multiple: e.resource.multiple,
+        value: e.resource.value,
+      }));
     }
   }
 
@@ -249,6 +251,8 @@ export default class AbilityConfigurationDialog extends PowerRollDialog {
 
     Hooks.off("targetToken", this.#targetHook);
 
+    this.#region.delete();
+
     return config;
   }
 
@@ -264,5 +268,26 @@ export default class AbilityConfigurationDialog extends PowerRollDialog {
     const { tokenUuid } = target.closest("[data-token-uuid]").dataset;
     const token = fromUuidSync(tokenUuid);
     await canvas.animatePan({ x: token.x, y: token.y });
+  }
+
+  /* -------------------------------------------------- */
+
+  /**
+   * The region currently placed by the ability. Replaced if place template is reused.
+   * @type {RegionDocument | null}
+   */
+  #region = null;
+
+  /* -------------------------------------------------- */
+
+  /**
+   * Place a template on the canvas and select all tokens inside.
+   * @this AbilityConfigurationDialog
+   * @param {PointerEvent} event   The originating click event.
+   * @param {HTMLElement} target   The capturing HTML element which defined a [data-action].
+   */
+  static async #placeTemplate(event, target) {
+    if (this.#region) await this.#region.delete();
+    this.#region = await this.item.system.placeTemplate({ setTargets: "replace" });
   }
 }