+++ src/module/data/item/ability.mjs
@@ -13,8 +13,8 @@ import enrichHTML from "../../utils/enrich-html.mjs";
  * @import { ApplicationConfiguration } from "@client/applications/_types.mjs";
  * @import { DatabaseCreateOperation } from "@common/abstract/_types.mjs";
  * @import RegionDocument from "@client/documents/region.mjs";
- * @import { RegionPlacementOptions } from "@client/canvas/layers/_types.mjs"
- * @import { PowerRollModifiers } from "../../_types.js";
+ * @import { PowerRollModifiers } from "../../_types";
+ * @import { PlaceAbilityOptions } from "./_types";
  * @import DrawSteelToken from "../../canvas/placeables/token.mjs";
  * @import DrawSteelTokenDocument from "../../documents/token.mjs";
  */
@@ -33,6 +33,7 @@ export default class AbilityModel extends BaseItemModel {
       detailsPartial: [systemPath("templates/sheets/item/partials/ability.hbs")],
       embedded: {
         PowerRollEffect: "system.power.effects",
+        SpecialEffect: "system.effects",
       },
     };
   }
@@ -97,10 +98,8 @@ export default class AbilityModel extends BaseItemModel {
       }, { persisted: false }),
     });
 
-    schema.effect = new fields.SchemaField({
-      before: new fields.HTMLField(),
-      after: new fields.HTMLField(),
-    });
+    schema.effects = new ds.data.fields.CollectionField(ds.data.pseudoDocuments.specialEffects.BaseSpecialEffect);
+
     schema.spend = new fields.SchemaField({
       value: new fields.NumberField({ integer: true }),
       text: new fields.StringField({ required: true }),
@@ -116,6 +115,50 @@ export default class AbilityModel extends BaseItemModel {
     // Game release updates
     if (data.type === "action") data.type = "main";
 
+    // 1.1 effect migration, based on Document._addDataFieldMigration
+    const { hasProperty, getProperty, setProperty, deleteProperty } = foundry.utils;
+    const { Document } = foundry.abstract;
+    if (!hasProperty(data, "effects")) {
+      const spendID = "spend".padEnd(16, "0");
+      if ((getProperty(data, "spend.value") || getProperty(data, "spend.text")) && !hasProperty(data, `effects.${spendID}`)) {
+        const value = getProperty(data, "spend.value");
+        setProperty(data, `effects.${spendID}`, {
+          _id: spendID,
+          type: "spend",
+          sort: CONST.SORT_INTEGER_DENSITY, // guarantees after "after" effects if present
+          resource: {
+            value,
+            multiple: value === null,
+          },
+          description: `<p>${getProperty(data, "spend.text") ?? ""}</p>`,
+        });
+        deleteProperty(data, "spend");
+      }
+      if (data.effect?.before) {
+        const beforeID = "before".padEnd(16, "0");
+        Document._addDataFieldMigration(data, "effect.before", `effects.${beforeID}`, (data) => {
+          const description = getProperty(data, "effect.before");
+          return {
+            _id: beforeID,
+            type: "base",
+            description,
+            before: true,
+          };
+        });
+      }
+      if (data.effect?.after) {
+        const afterID = "after".padEnd(16, "0");
+        Document._addDataFieldMigration(data, "effect.after", `effects.${afterID}`, (data) => {
+          const description = getProperty(data, "effect.after");
+          return {
+            _id: afterID,
+            type: "base",
+            description,
+          };
+        });
+      }
+    }
+
     return super.migrateData(data);
   }
 
@@ -124,6 +167,8 @@ export default class AbilityModel extends BaseItemModel {
   /** @inheritdoc */
   prepareDerivedData() {
     super.prepareDerivedData();
+
+    this.#resourceName = null;
 
     this.power.roll.enabled = !this.power.roll.reactive && (this.power.effects.size > 0);
   }
@@ -309,6 +354,7 @@ export default class AbilityModel extends BaseItemModel {
 
     const targetConfig = ds.CONFIG.abilities.targets[this.target.type] ?? { embedLabel: "COMMON.Unknown" };
     if (this.target.custom) labels.target = this.target.custom;
+    // == null covers null & undefined
     else if (this.target.value === null) labels.target = targetConfig.all;
     if (!labels.target) {
       // Non-plural dependent labels
@@ -324,26 +370,40 @@ export default class AbilityModel extends BaseItemModel {
 
   /* -------------------------------------------------- */
 
-  /** @inheritdoc */
-  async getSheetContext(context) {
-    const config = ds.CONFIG.abilities;
-    const formattedLabels = this.formattedLabels;
+  /**
+   * Cached reference to the resource name, reset during data prep.
+   * @type {string}
+   */
+  #resourceName = null;
 
-    let resourceName = this.actor?.system.coreResource?.name;
+  /**
+   * A cached reference to the heroic resource consumed by this ability.
+   * @type {string}
+   */
+  get resourceName() {
+    this.#resourceName ??= this.actor?.system.coreResource?.name;
 
-    if (!resourceName && (this.prerequisites.dsid.size === 1)) {
+    if (!this.#resourceName && (this.prerequisites.dsid.size === 1)) {
       const dsid = this.prerequisites.dsid.first();
       let classEntry = ds.registry.class.filter(e => e.dsid === dsid).at(-1);
       if (!classEntry) {
         const subclass = ds.registry.subclass.filter(e => e.dsid === dsid).at(-1);
         if (subclass) classEntry = ds.registry.class.filter(e => e.dsid === subclass.classLink).at(-1);
       }
-      if (classEntry) resourceName = classEntry.primary;
+      if (classEntry) this.#resourceName = classEntry.primary;
     }
 
-    resourceName ??= _loc("DRAW_STEEL.Actor.hero.FIELDS.hero.primary.value.label");
+    return this.#resourceName ??= _loc("DRAW_STEEL.Actor.hero.FIELDS.hero.primary.value.label");
+  }
 
-    context.resourceName = resourceName;
+  /* -------------------------------------------------- */
+
+  /** @inheritdoc */
+  async getSheetContext(context) {
+    const config = ds.CONFIG.abilities;
+    const formattedLabels = this.formattedLabels;
+
+    context.resourceName = this.resourceName;
 
     context.keywordList = formattedLabels.keywords;
 
@@ -387,13 +447,15 @@ export default class AbilityModel extends BaseItemModel {
       context.powerRollBonus = this.power.roll.formula.replace("@chr", characteristicsFormatter.format(Array.from(characteristicList)));
     }
 
-    context.enrichedBeforeEffect = await enrichHTML(this.effect.before, { relativeTo: this.parent });
-    context.enrichedAfterEffect = await enrichHTML(this.effect.after, { relativeTo: this.parent });
-
-    context.spendLabel = _loc("DRAW_STEEL.Item.ability.SpendLabel", {
-      value: this.spend.value ?? "",
-      name: resourceName,
-    });
+    context.beforeEffects = [];
+    context.afterEffects = [];
+    for (const effect of this.effects.sortedContents) {
+      const displayData = {
+        label: effect.label,
+        text: await enrichHTML(effect.description, { relativeTo: this.parent }),
+      };
+      context[effect.before ? "beforeEffects" : "afterEffects"].push(displayData);
+    }
   }
 
   /* -------------------------------------------------- */
@@ -544,9 +606,7 @@ export default class AbilityModel extends BaseItemModel {
 
     let resourceSpend = fd.resource ?? 0;
 
-    if (fd.spend) {
-      resourceSpend += typeof fd.spend === "boolean" ? this.spend.value : fd.spend;
-    }
+    for (const spend of Object.values(fd.spend ?? {})) resourceSpend += spend;
 
     if (resourceSpend) {
       messageData.flavor = _loc("DRAW_STEEL.Item.ability.ConfigureUse.SpentFlavor", {
@@ -676,14 +736,14 @@ export default class AbilityModel extends BaseItemModel {
 
   /**
    * Create a region template based on this ability's distance data.
-   * @param {RegionPlacementOptions} [options={}] Options to forward to canvas.regions.placeRegion.
+   * @param {PlaceAbilityOptions} [options={}] Options to forward to canvas.regions.placeRegion.
    * @returns {Promise<RegionDocument>} The Region document that was placed or null if
    *  - the placements of all shapes were skipped,
    *  - the dismiss key was pressed,
    *  - the game is paused and the user is not a GM, or
    *  - the Region creation was rejected by preCreate.
    */
-  async placeTemplate(options = {}) {
+  async placeTemplate({ setTargets, ...options } = {}) {
     if (!this.hasTemplate) {
       const msg = _loc("DRAW_STEEL.Item.ability.NoArea", { ability: this.parent.name });
       ui.notifications.error(msg, { console: false });
@@ -746,6 +806,40 @@ export default class AbilityModel extends BaseItemModel {
       },
     };
 
-    return canvas.regions.placeRegion(regionData, options);
+    const region = await canvas.regions.placeRegion(regionData, options);
+    if (!setTargets || !region) return region;
+
+    /** @type {Set<DrawSteelToken>} */
+    const tokens = canvas.tokens.quadtree.getObjects(region.bounds);
+    const targetIds = tokens.filter(token => {
+      return this.validTarget(token.document) && token.document.testInsideRegion(region);
+    }).map(token => token.id);
+    if (targetIds.size) canvas.tokens.setTargets(targetIds, { mode: setTargets });
+    return region;
+  }
+
+  /* -------------------------------------------------- */
+
+  /**
+   * Validates a given token against this ability's.
+   * @param {DrawSteelTokenDocument} token The token to potentially target.
+   * @returns {boolean}
+   */
+  validTarget(token) {
+    const targetOptions = ds.CONFIG.abilities.targets[this.target.type]?.targetOptions;
+    if (!targetOptions) return false;
+
+    if (token.actor.uuid === this.actor.uuid) return targetOptions.has("self");
+    if (token.actor.system.isObject && targetOptions.has("object")) return true;
+
+    if (token.actor.system.isCreature) {
+      const sourceDisposition = this.actor.getDependentTokens({ scenes: canvas.scene })?.at(0)?.disposition
+        ?? this.actor.prototypeToken.disposition;
+      const polarization = [CONST.TOKEN_DISPOSITIONS.FRIENDLY, CONST.TOKEN_DISPOSITIONS.HOSTILE];
+      if (targetOptions.has("ally") && targetOptions.has("enemy")) return true;
+      else if (targetOptions.has("ally")) return sourceDisposition === token.disposition;
+      else if (targetOptions.has("enemy")) return (sourceDisposition !== token.disposition) && polarization.includes(sourceDisposition);
+    }
+    return false;
   }
 }